добавлена возможность постановки лайка

+++ src/utils/Api.js
@@ -82,27 +82,30 @@ class Api {
             .then(res => this._getResponse(res))
     }
 
-    // поставить лайк
-    putLikeOnCard(cardId) {
-        return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-            method: 'PUT',
-            headers: {
-                authorization: this._token,
-            },
-        })
-            .then(res => this._getResponse(res))
+    changeLikeCardStatus(cardId, isLiked) {
+        // елси карточка лайкнута
+        if (isLiked) {
+            // запрос на удаление лайка
+            return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+                method: 'DELETE',
+                headers: {
+                    authorization: this._token,
+                },
+            })
+                .then(res => this._getResponse(res))
+        // иначе
+        } else {
+            // запрос на постановку лайка
+            return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+                method: 'PUT',
+                headers: {
+                    authorization: this._token,
+                },
+            })
+                .then(res => this._getResponse(res))
+        }
     }
 
-    // удалить лайк
-    deleteLikeFromCard(cardId) {
-        return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-            method: 'DELETE',
-            headers: {
-                authorization: this._token,
-            },
-        })
-            .then(res => this._getResponse(res))
-    }
 
     _getResponse(res) {
         if (res.ok) {
